requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.22.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.75.0
anyio==4.12.0
asn1crypto==1.5.1
asyncio==4.0.0
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
boto3==1.42.15
botocore==1.42.15
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
distro==1.9.0
docstring_parser==0.17.0
ecdsa==0.19.1
fastapi==0.128.0
filelock==3.20.2
Flask==3.1.2
fpdf==1.7.2
frozenlist==1.8.0
fsspec==2025.12.0
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.45.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.3
google-genai==1.56.0
googleapis-common-protos==1.72.0
googlesearch-python==1.3.0
graphviz==0.21
greenlet==3.3.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.2.4
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
kafka-python==2.3.0
langchain==1.2.0
langchain-core==1.2.5
langchain-openai==1.1.6
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.1
langsmith==0.5.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.7.0
networkx==3.6.1
numpy @ file:///home/builder/.termux-build/python-numpy/src/dist/numpy-2.2.5-cp312-cp312-linux_aarch64.whl#sha256=5bd061eb5e1323ce3e3f3e293dd720cbdff3a5b3ac066b0fcc963cb32949730d
oauthlib==3.3.1
ollama==0.6.1
openai==2.14.0
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
pandas @ file:///home/builder/.termux-build/python-pandas/src/dist/pandas-2.3.3-cp312-cp312-linux_aarch64.whl#sha256=d69521bd9c24058a0abad646143cca00bde1cbc8f9852e7a5d077d9abd359db9
perplexityai==0.22.3
pg8000==1.31.5
pillow==12.0.0
pluggy==1.6.0
propcache==0.4.1
proto-plus==1.27.0
protobuf==6.33.2
psutil==7.2.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.1
pytest==9.0.2
pytest-timeout==2.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rsa==4.9.1
s3transfer==0.16.0
scramp==1.4.6
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.1
SQLAlchemy==2.0.45
starlette==0.50.0
stripe==14.1.0
tenacity==9.1.2
termuxgui==0.1.6
tiktoken==0.12.0
tqdm==4.67.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.2
uuid_utils==0.12.0
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
Werkzeug==3.1.4
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
